backend/requirements.txt
Django==2.2.5
google-cloud-vision==0.39.0